// src/frontend/react-admin/src/generated/models/avis-request.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Tracking Application
 * Backend APIs pour une application de Tacking Visites
 *
 * OpenAPI spec version: 1.0
 * Contact: donfackduval@gmail.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { VisiteResponse } from './visite-response';
/**
 * 
 * @export
 * @interface AvisRequest
 */
export interface AvisRequest {
    /**
     * 
     * @type {string}
     * @memberof AvisRequest
     */
    libelle?: string;
    /**
     * 
     * @type {VisiteResponse}
     * @memberof AvisRequest
     */
    visite?: VisiteResponse;
}