// src/frontend/react-admin/src/generated/models/avis-response.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Tracking Application
 * Backend APIs pour une application de Tacking Visites
 *
 * OpenAPI spec version: 1.0
 * Contact: donfackduval@gmail.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { VisiteResponse } from './visite-response';
/**
 * 
 * @export
 * @interface AvisResponse
 */
export interface AvisResponse {
    /**
     * 
     * @type {number}
     * @memberof AvisResponse
     */
    id?: number;
    /**
     * 
     * @type {string}
     * @memberof AvisResponse
     */
    libelle?: string;
    /**
     * 
     * @type {VisiteResponse}
     * @memberof AvisResponse
     */
    visite?: VisiteResponse;
    /**
     * 
     * @type {Date}
     * @memberof AvisResponse
     */
    createdAt?: Date;
    /**
     * 
     * @type {Date}
     * @memberof AvisResponse
     */
    updatedAt?: Date;
}